[ADD] add food dainamicakly

+++ pages/api/products/[id].js
@@ -8,7 +8,7 @@ export default async function handler(req, res) {
     query: { id },
     cookies
   } = req;
-  // const token = cookies.token
+  const token = cookies.token
 
   dbConnect();
 
@@ -22,9 +22,9 @@ export default async function handler(req, res) {
   }
 
   if (method === "PUT") {
-    // if(!token || token !== process.env.token){
-    //   return res.status(401).json("Not authenticated!")
-    // }
+    if(!token || token !== process.env.token){
+      return res.status(401).json("Not authenticated!")
+    }
     try {
       const product = await Product.findByIdAndUpdate(id, req.body, {
         new: true,
@@ -36,9 +36,9 @@ export default async function handler(req, res) {
   }
 
   if (method === "DELETE") {
-    // if(!token || token !== process.env.token){
-    //   return res.status(401).json("Not authenticated!")
-    // }
+    if(!token || token !== process.env.token){
+      return res.status(401).json("Not authenticated!")
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("The product has been deleted!");
